pyproject.toml
[project]
name = "micro-app-mcp"
version = "0.1.0"
description = "MCP Server for micro-app knowledge base"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.1.0,<2.0",
    "fastapi>=0.109.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "chromadb>=0.4.0",
    "playwright>=1.40.0",
    "huggingface-hub>=0.20.0",
    "sentence-transformers>=2.0.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "PyGithub>=1.58.0",
    "mcp>=1.26.0",
    "html2text>=2020.1.16",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.23.0"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" } # 告诉 setuptools：根包在 src/ 下，决定 wheel 包的目录结构没有包含 src 目录本身

[tool.setuptools.packages.find]
where = ["src"]             # 在 src/ 目录下查找包，有配置 package-dir，此处可省略
include = ["micro_app_mcp"] # 只包含 micro_app_mcp 包

[project.scripts]
micro-app-mcp = "micro_app_mcp.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"         # 自动检测异步测试
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
